done the pay on counter
